UI main kodo skaidymas
